Update dogs_age.py
To calculate the dogs age in human years.  The calculation is if the dog is 2 then it is 21 years and for every year after that age is x4.  I googled this one as I hadnt even started learning python let alone write a program.  The youtube video I watched on this was too complex for being a simple equation so I ended up cobbling it together instinctively.  It did not work because the float for the human age was not properly defined so I had some help from another student to correctly define the float for the human age.  It worked! but alas the calculation was incorrect because I did not subtract the 2 years from the dogs age.  Corrected this and started to understand the if and else statements.